+++ DOC/user/models.py
@@ -50,9 +50,10 @@ class Profile(models.Model):
         return str(self.user)
 
     def delete(self, *args, **kwargs):
-        # You have to prepare what you need before delete the model
-        storage, path = self.profile_image.storage, self.profile_image.path
-        # Delete the model before the file
-        super(Profile, self).delete(*args, **kwargs)
-        # Delete the file after the model
-        storage.delete(path)
+        if self.profile_image:
+            # You have to prepare what you need before delete the model
+            storage, path = self.profile_image.storage, self.profile_image.path
+            # Delete the model before the file
+            super(Profile, self).delete(*args, **kwargs)
+            # Delete the file after the model
+            storage.delete(path)
